forbid some actions for readonly superadmins serverside

+++ back/server/api/delete_allowed_email.ts
@@ -3,6 +3,9 @@ import { logError } from '../helpers/logger';
 
 export const delete_allowed_email = async (req: any, res: any): Promise<void> => {
   try {
+    if (req.session.isReadOnlySuperadmin) {
+      return res.status(401).end();
+    }
     const allowedEmailId = req.params.id;
     await db.query(`DELETE FROM allowed_emails WHERE id=$1 AND group_id=$2`, [
       allowedEmailId,
@@ -10,7 +13,7 @@ export const delete_allowed_email = async (req: any, res: any): Promise<void> =>
     ]);
     res.status(200).end();
   } catch (e) {
-    logError("delete_allowed_email", e);
+    logError('delete_allowed_email', e);
     res.status(400).end();
   }
 };
