+++ back/server/api/delete_device.ts
@@ -3,6 +3,9 @@ import { logError } from '../helpers/logger';
 
 export const delete_device = async (req: any, res: any): Promise<void> => {
   try {
+    if (req.session.isReadOnlySuperadmin) {
+      return res.status(401).end();
+    }
     const deviceId = req.params.deviceId;
     await db.query('DELETE FROM user_devices WHERE id=$1 AND group_id=$2', [
       deviceId,
