+++ back/server/api/delete_group_admin.ts
@@ -3,6 +3,9 @@ import { logError } from '../helpers/logger';
 
 export const delete_group_admin = async (req: any, res: any): Promise<void> => {
   try {
+    if (req.session.isReadOnlySuperadmin) {
+      return res.status(401).end();
+    }
     const adminId = req.params.id;
     const deletedAdmin = await db.query(
       `DELETE FROM admin_groups WHERE admin_id=$1 AND group_id=$2`,
@@ -17,7 +20,7 @@ export const delete_group_admin = async (req: any, res: any): Promise<void> => {
     });
     res.status(200).end();
   } catch (e) {
-    logError("delete_group_admin", e);
+    logError('delete_group_admin', e);
     res.status(400).end();
   }
 };
